requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
praat-parselmouth==0.4.3
openpyxl==3.1.2
werkzeug==3.0.1
